+++ src/app/app.module.ts
@@ -1,7 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { ProductModule, CartModule, CoreModule } from './modules';
 
+import { ProductModule, CartModule, CoreModule, DemoModule } from './modules';
 import { AppComponent } from './app.component';
 
 @NgModule({
@@ -12,7 +12,8 @@ import { AppComponent } from './app.component';
     BrowserModule,
     CoreModule,
     ProductModule,
-    CartModule
+    CartModule,
+    DemoModule
   ],
   providers: [],
   bootstrap: [AppComponent]