+++ src/app/modules/core/components/index.ts
@@ -1,1 +1,1 @@
-export * from './core-demo/core-demo.component';
+// export * from './core-demo/core-demo.component';