+++ src/app/modules/core/pipes/category-display/category-display.pipe.spec.ts
@@ -1,8 +1,0 @@
-import { CategoryDisplayPipe } from './category-display.pipe';
-
-describe('CategoryDisplayPipe', () => {
-  it('create an instance', () => {
-    const pipe = new CategoryDisplayPipe();
-    expect(pipe).toBeTruthy();
-  });
-});