+++ src/app/modules/core/pipes/index.ts
@@ -1,1 +1,0 @@
-export * from './category-display/category-display.pipe';