+++ src/app/modules/demo/components/index.ts
@@ -1,0 +1,1 @@
+export * from './core-demo/core-demo.component';