+++ src/app/modules/index.ts
@@ -1,3 +1,4 @@
 export * from './product/product.module';
 export * from './cart/cart.module';
 export * from './core/core.module';
+export * from './demo/demo.module';