+++ src/app/services/local-storage/local-storage.service.ts
@@ -23,4 +23,9 @@ export class LocalStorageService {
       localStorage.removeItem(key);
     }
   }
+  clearStorage() {
+    if (this.available) {
+      localStorage.clear();
+    }
+  }
 }