+++ changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 1.4.0 - 2018-05-31
+### Added
+
+### Changed
+- Core Demo component moved to Demo module
+- Single product query (via promise)
+
 ## 1.3.0 - 2018-05-29
 ### Added
 - Category Display pipe